gss: Fix --limit-enctype test
This is a bug that was introduced by 5447b81fb1c6ec72a4dede5284bf21edd98ba65d
(which made it so that on-store we derive an `etypes` for every HDB entry that
had that field absent), but masked by a bug in `_kdc_find_etype()`.

The fix is to not just rely on `kadmin del_enctype` for affecting ticket
session key negotiation, but to also or instead use `kadmin mod -e`.