kadm5: Add exec_hook built-in hook plugin
You can now have [kadmin] exec_hook = path-to-program in krb5.conf and
kadmin -l, kadmind, and httpkadmind, and any user of libkadm5srv will
end up executing that hook.

That will execute the hook program for every kadmin operation.  This is
done via a built-in exec hook plugin for the kadm5 plugin interface.

Exiting with an error during a pre-commit execution of the hook will
prevent the operation from being committed.

For password changes via kadmin, this can also be used as a separate
password quality check hook, though this hook can only stop the change
but without any explanation.