hdb: HDB iteration and dump improvements
We introduce an `hdb_foreach_eoa()` function whose callback will get the
key (principal name) as well as the value, and which will get both,
entries and aliases.

Not all backends support providing the key, so it may be NULL.

As well, we add support for Heimdal-JSON and Heimdal-JSON-nokeys format
types for HDB dumping.