asn1: Fix templating of inlined CHOICEs
We use an inlined CHOICE type in lib/hdb/hdb.asn1 as the type of the
`data` member of `HDB-extension`.

The template backend emits `offsetof(struct HDB_extension, data.element)`
and `offsetof(struct HDB_extension, data.u.pkinit_acl)` and so on for
the offsets of the relevant fields of that CHOICE's corresponding C
struct, but the "parent" template entry for the `data` member of
`HDB-extension` would also say `offsetof(struct HDB_extension, data)`,
thus the offset of `data` in `struct HDB_extension` would get added
twice.

This commit has one fix for this which is to set the offset in the
parent entry to 0, but really we should set the offsets in the CHOICE
type's template members to offsets off of HDB_extension_data -- we need
to know that the type is anonymous so that we can append the member
name.  Still, for now this will do.