hdb: Switch to ASN.1 compiler template backend
The ASN.1 compiler's template backend currently does not handle SEQUENCE
members of inlined CHOICE type: the offsets encoded into the CHOICE's
template elements are off by the offset of the member of that type in
its SEQUENCE.  Therefore we're promoting the HDB-extension data member's
CHOICE type to a top-level type.  We'll fix this later, but probably
leave this un-inlining.

A consequence of this is that all the choice enum value symbol names for
extensions changed.