asn1: For now clamp GeneralizedTime to 2038
Sadly, krb5_timestamp is an krb5_int32, even on 64-bit platofmrs.  That
means that even on 64-bit we have a year 2038 problem.

We'll have to make krb5_timestamp unsigned, but it will still be a
32-bit int.

For now we want to clamp any KerberosTime (GeneralizedTime) values from
any PDUs to 2^31-1.  We do this by just clamping GeneralizedTime in
lib/asn1.

A later commit will make sure that we always clamp KerberosTime values
in Kerberos code, then we can change the ASN.1 code touched in this
commit to clamp datetimes past 2106.