krb5: WIP clamp KerberosTime to 2038
Sadly, krb5_timestamp is an krb5_int32, even on 64-bit platforms.  That
means that even on 64-bit we have a year 2038 problem.

We're going to have to make krb5_timestamp an unsigned 32-bit int.  We
can change its signedness, but not its size.  Even then, we'll have a
2106 problem.

For now we want to clamp any KerberosTime (GeneralizedTime) values from
any PDUs to 2^31-1.

This commit currently just renames all KerberosTime fields in Kerberos
ASN.1 types so that it's easy to identify all of them.  Next up we're
going to add a macro to every reference to these to clamp any
KerberosTime values from PDUs to 2^31-1.  That's why this commit is a
WIP: we still have to do that.

When we're done clamping every reference to KerberosTime values from
PDUs we can undo the KerberosTime field renaming in this commit, and
then it will stop being a WIP commit.  Though leaving the fields renamed
might well prove useful later, so maybe we shouldn't rename them back?