kuser: kimpersonate with HDBGET: keytab
With this change, if [krb5] plugin_dir is set to a directory that
contains hdb_keytab_plugin.so then kimpersonate can use the HDBGET:
keytab type.

This is useful for KDC-side user impersonation, which could be
useful for low infrastructure break-glass access.  Low infrastructure
because the host need only run ipropd-slave, and even if that's down
the local HDB can be used for impersonation.