krb5: Ignore other collections when KRB5CCNAME is set (fix #1124)
When KRB5CCNAME was set (or a default ccache was set via
krb5_cc_set_default_name()) don't search any other collections in
krb5_cc_cache_match() or krb5_cccol_last_change_time().

Otherwise one could `kinit -c $SOME_NON_DEFAULT_CCACHE ...` and have
kinit look for and find a suitable ccache in some other collection.

The point of searching all collections is that if we know they're all
relevant in the current context, then we can save the user (or login
apps) having to set and coordinate KRB5CCNAME, which would be a
significant UI/UX improvement.  But we can only do that when we know
that a ccache [collection] was not named either via KRB5CCNAME nor more
explicitly.

Well, _if_ the ccache named via KRB5CCNAME or otherwise happens to be
the same as the default ccache for that cache type, then perhaps we
should still search the other cache types' default collections too, but
that will require more work.