com_err: Install a heim_compile_et
In order to use `./configure --host=... --with-tools=DIR ...` one needs a
directory in which to find: `slc`, `compile_et`, and `asn1_compile`.

`slc` already gets installed into `${libexecdir}/`, and `asn1_compile` gets
installed into `${bindir}/`, but `compile_et` does not and should not get
installed because it would conflict with the system's com_err.  So let's
install a `heim_compile_et` into `${bindir}`.