kdc: Add KDC support for PKINIT Freshness extension (RFC 8070)
Clients indicating support for PKINIT Freshness (by means of an empty
PADATA_AS_FRESHNESS type) will receive in the reply a freshness token to
be included in the client’s PK-INIT request which proves the client’s
recent possession of the private key.

The require_pkinit_freshness option, if enabled, will reject PK-INIT
requests that lack this token.

Freshness tokens are only supported with the PA-PK-AS-REQ padata type.
The PA-PK-AS-REQ-Win2k type has no field to contain a freshness token.

Signed-off-by: Jennifer Sutton <jennifersutton@catalyst.net.nz>