krb5: Fix Java interop problem in krb5_mk_error_ext()
krb5_mk_error_ext() with a NULL server name would generate a KRB-ERROR
with its sname field set to a zero-component principal name.  This
causes Java's Krb5 client to throw an IllegalArgumentException with
`Empty nameStrings not allowed` as its message.

Now krb5_mk_error_ext() with a NULL server name will set the sname field
to a one-component name with the one component being "<>".  We also
minimize the srealm, setting it to "<>" as well.